Move minOps to server
