+++ server/src/main/java/module-info.java
@@ -480,4 +480,5 @@ module org.elasticsearch.server {
     exports org.elasticsearch.inference.configuration;
     exports org.elasticsearch.monitor.metrics;
     exports org.elasticsearch.plugins.internal.rewriter to org.elasticsearch.inference;
+    exports org.elasticsearch.lucene.util.automaton;
 }
