+++ server/src/test/java/org/elasticsearch/lucene/util/automaton/MinimizationOperationsTests.java
@@ -1,20 +1,21 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-package org.elasticsearch.xpack.core.security.support;
+package org.elasticsearch.lucene.util.automaton;
 
 import org.apache.lucene.tests.util.automaton.AutomatonTestUtil;
 import org.apache.lucene.util.automaton.Automaton;
 import org.apache.lucene.util.automaton.Operations;
 import org.elasticsearch.test.ESTestCase;
 
-import static org.elasticsearch.xpack.core.security.support.Automatons.subsetOf;
+import static org.apache.lucene.tests.util.automaton.AutomatonTestUtil.subsetOf;
 
-// partially copied from Lucene's AutomatonTestUtil
 public class MinimizationOperationsTests extends ESTestCase {
 
     /** the minimal and non-minimal are compared to ensure they are the same. */
